PMM-6279 Show query metadata in the details
