+++ pmm-app/src/pmm-qan/panel/components/Details/Metadata/Metadata.types.ts
@@ -1,0 +1,6 @@
+// Interface for the line in metadata table
+export interface LineMetadata {
+  name: string;
+  value: string;
+  tooltip: string;
+}
